Update sound.coffee
Remove all HTML5 audio tag code in favor of AudioContext API. This greatly reduces the complexity of the Sound class and is potentially far more powerful. Testing is required as soon as possible, and if all goes well support of different audio nodes is the next major goal as far as the Sound class is concerned.