add start and stop functionality
In order to properly start and stop the engine, the user should activate
an event. In this case I've simply added an onclick method to the canvas.
The first time it is clicked, all sounds will silently play for 50
milliseconds to preload them (a hack to get around apple's safari policy).
Also all sprite _update functions and sound play functions
will start and stop with the engine. This brings the whole engine
together and is much more natural than the old setup.