seperate style from library
This commit solves issue #1 by remove all style code from the library
and placing it in src/style.less. A new task has been added to the Cakefile
to allow for automatic building of lib/GreenhornStyle.css using the
command line tool lessc. The purpose of this commit is to make it easy
for a user to define their own style using less or css without having
to hack the library, all the while using a template they can be sure works.