perfect circleOfFriends.coffee
I have been testing and modifying and testing and modifying and testing
and modifying for a few days, and I believe I have finally settled on
an official style. All that left to do is add/remove/change the other
examples to perfectly exemplify what this engine is made of.