address issues #14 and #15
I've made a few updates to the Sound API. Firstly, _config.playOnLoad
has been changed to _config.autoplay for the reasons listed in issue #14.
Secondly, I've added pause functionality so that the user can resume a
particular sound and not have to start all over. And lastly, no sound
operations will be performed unless the engine is currently running.
I've also addressed issue #15 by appending a '/' character to the end
of environment PATHs when they are used if the user forgot to include one.