rework Sprite class to fire events
All the events listed in issue #5 have been rougly implemented, except
the mouse events. I am still unsure how I want to handle those. Other
changes include: not checking for visibility in the _update functions,
TextSprite and AniSprite now call Sprite._update first to ensure that
the on('update') event always fires before anything else, and vector
magnitudes and angles are seperated from the Sprite construction config
object so that they can be set in the proper order. I still need to: write
a factory function that returns proper callbacks for boundary collisions,
determine if on('get'), on('set'), and on('change') are a good idea, and
determine what events should be fired from the other classes, primarily
the Greenhorn class and how it relates to issue #11.