finalize Sprite events
All of the Sprite events that I wanted to implement are now in place.
The mouse events are fired from the '#gh-canvas' element to all Sprites.
The boundary action callbacks required the creation of another primary
object: _bas. These events still need to be tested, but the next commit
will likely address issue #11, which relates to Game State.